fix(stm32f401): address review comments on NUCLEO-F401RE support
Required:
- Add APB1_FREQ/APB2_FREQ and use them for SPI, I2C, and UART clock divisors
- Move getBaudRateDivisor into the shared F4 file
- Fix F4 ADC sample-time shifts and drop machine_stm32f401_adc.go
- Add ADC0-ADC15 aliases and examples/pwm/nucleo-f401re.go

Optional:
- Use a bare stm32f401 build tag on machine_stm32f401.go
- Switch the full-peripheral F4 file to a positive F405/F407/F469 tag
- Rename machine_stm32f4_extended.go to machine_stm32f40x_periph.go
- Enable HSE with SetBits and set APB prescalers before switching to PLL
- Reindent targets/nucleo-f401re.json to 2 spaces
